minor fix and change logic for take a specific disservice
