+++ src/main/java/it/pn/frontend/e2e/utility/DataPopulation.java
@@ -4,12 +4,14 @@ import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.dataformat.yaml.YAMLFactory;
 import com.fasterxml.jackson.dataformat.yaml.YAMLGenerator;
+import lombok.Setter;
 import org.junit.Assert;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.File;
 import java.io.IOException;
+import java.time.ZoneId;
 import java.util.HashMap;
 import java.util.Map;
 
@@ -17,6 +19,26 @@ public class DataPopulation {
     private static final Logger logger = LoggerFactory.getLogger("DataPopulation");
 
     private ObjectMapper objectMapper;
+
+    @Setter
+    private String dataDa;
+    @Setter
+    private String dataA;
+    private ZoneId targetZoneId; // zona target, puoi cambiarla a tuo piacimento
+
+    public DataPopulation() {
+        this.targetZoneId = ZoneId.of("Europe/Paris");
+    }
+
+    public String getDataA() {
+        String dataFormatted = dataA.replace("ore", "").trim();
+        return WebTool.convertToLocalTime(dataFormatted, targetZoneId);
+    }
+
+    public String getDataDa() {
+        String dataFormatted = dataDa.replace("ore", "").trim();
+        return WebTool.convertToLocalTime(dataFormatted, targetZoneId);
+    }
 
     public Map<String, Object> readDataPopulation(String dpFile) {
         this.objectMapper = new ObjectMapper(new YAMLFactory());
