+++ src/test/resources/feature/2-mittente/6_disserviziAppPA/025_A25_downloadAttestazioniDisserviziRisolti.feature
@@ -18,7 +18,7 @@ Feature: Mittente effettua il download delle attestazioni dei disservizi risolti
     When Nella pagina Piattaforma Notifiche selezionare la voce 'stato della piattaforma'
     Then Si visualizza correttamente la pagina dello 'stato della piattaforma' di mittente
     And Si visualizza correttamente la tabella dei disservizi
-    And Si visualizza un record in elenco relativo ad un disservizio risolto
+    And Si visualizza un record in elenco relativo ad un disservizio risolto "Invio delle notifiche"
     And Download file attestazione disservizio
     And Controllo corrispondenza dati con pdf
     And Logout da portale mittente
